+++ Cargo.cef.lock
@@ -34,7 +34,7 @@ dependencies = [
  "pdf",
  "pem 4.0.0",
  "percent-encoding",
- "quick-xml 0.36.2",
+ "quick-xml 0.41.0",
  "rand 0.8.7",
  "read-progress-stream",
  "reqwest 0.12.28",
@@ -6278,15 +6278,6 @@ name = "quick-error"
 version = "2.0.1"
 source = "registry+https://github.com/rust-lang/crates.io-index"
 checksum = "a993555f31e5a609f617c12db6250dedcac1b0a85076912c436e6fc9b2c8e6a3"
-
-[[package]]
-name = "quick-xml"
-version = "0.36.2"
-source = "registry+https://github.com/rust-lang/crates.io-index"
-checksum = "f7649a7b4df05aed9ea7ec6f628c67c9953a43869b8bc50929569b2999d443fe"
-dependencies = [
- "memchr",
-]
 
 [[package]]
 name = "quick-xml"
